[Enterprise Search] Move schema types to shared (#84822)
* Move schema types to shared

We use the Schema types in Workplace Search as well, so moving these to shared. Also, we have a component called IndexingStatus so reverting to the prefixed IIndexingStatus interface name

* Fix misspelled interface
